ollama_fastapi_docker/fastapi/requirements.txt
aiohappyeyeballs==2.4.0
Requests
chromadb
fastapi==0.110.2
uvicorn==0.29.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
fastembed==0.3.6
filelock==3.15.4
flash==1.0.3
Flask==3.0.3
Flask-Cors
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.112
loguru==0.7.2
MarkupSafe==2.1.5
marshmallow==3.22.0
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.19.2
orjson==3.10.7
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
pydantic==2.8.2
pydantic_core==2.20.1
pyreadline3==3.4.1
PyStemmer==2.2.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.34
sympy==1.13.2
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
win32-setctime==1.1.0
yarl==1.9.9
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.1.2
build==1.2.1
cachetools==5.3.3
cffi==1.16.0
chroma-hnswlib==0.7.3
coloredlogs==15.0.1
cryptography==42.0.5
dataclasses-json==0.6.7
Deprecated==1.2.14
fastapi==0.110.2
fastembed==0.3.6
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
importlib-metadata==7.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==29.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.112
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.2
onnxruntime==1.19.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.4.0
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
pypdfium2==4.29.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.37.2
sympy==1.13.2
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.9.9
zipp==3.18.1
python-multipart